+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 0.3.0 - 19 November 2023
+### Added
+- Overhauled the message multiplier system to be more flexible and hopefully capable of producing better results
+- New setting `max_messages`: This is the maximum number of recent messages that will be injected into the negative prompt
+- New setting `max_multiplier`: This is the maximum multiplier that will be applied to a message (in other words, the repeat count)
+- New setting `scaling`: This is the scaling algorithm that will be applied to the multiplier for each message, currently supports `constant`, `linear`, `exponential` and `logarithmic`
+
+### Removed
+- `history_multiplier` setting
+- `last_message_multiplier` setting
+
 ## 0.2.0 - 18 November 2023
 ### Added
 - Default tab support